rename: Generator 파일 service 패키지로 이동
