+++ src/main/java/bridge/service/PlayerSetting.java
@@ -1,12 +1,10 @@
 package bridge.service;
 
-import bridge.BridgeRandomNumberGenerator;
 import bridge.domain.Player;
 import bridge.dto.BridgeCorrectLocationDto;
 import bridge.dto.BridgeDto;
 import bridge.dto.PlayerBridgeSizeDto;
 import bridge.exception.InputException;
-import bridge.service.BridgeMaker;
 
 import java.util.List;
